feat: Added APIKIT_VERSION rule
There is a memory leak issue in APIKit versions <= 1.3.11 (https://help.mulesoft.com/s/article/Out-Of-Memory-Metaspace-while-often-re-deploying-applications-using-APIKit).

Implemented APIKIT_VERSION rule to check the version of APIKit dependency in pom.xml is at least artifactVersion passed in the rule.

artifactVersion for APIKit dependency is defaulted to '1.9.0', and this is configurable in the rule.