Updates base POM rules to use pom.xml as filename
With the feature to support parent POM by evaluating the effective pom,
the base POM rules where reporting validation errors on the effective
pom file instead of the pom.xml.
This would cause SonarQube to ignore these errors, as they are reported
on a file that does not exist in the code base.